requirements.txt
requests>=2.32.4
PyYAML>=6.0.2
yt-dlp[default]>=2026.03.17
flask>=3.0.0